feat: Add layer blending mode and new layer functionality
Introduces a `blendingMode` property to the `Layer` type and updates the `LayersPanel` to allow users to select blending modes. Also adds a new "Add Layer" button in the `LayersPanel` which calls a new `addNewLayer` action in the store. The `useEffect` hook for zoom controls has been removed from `EditorPage` as it was not being used. The canvas background drawing logic in `CanvasComponent` has been simplified to allow for checkerboard patterns to show through transparent backgrounds.